refactor: 디버깅용 console 문 전체 제거 (47개 파일, 181개)
- hooks/dragAndDrop/useDragDrop.ts (31개), useLayoutMeasurement.ts (30개)
- app/(tabs)/explore/detail/[id].tsx (20개), user/flight-detail/[id].tsx (20개)
- store/useAuthStore.ts (7개), flightLogStore.ts (5개), locationStore.ts (2개)
- libs/ 하위 fetch 함수 14개 파일
- api/apiClient.ts 응답 로깅 제거
