+++ app/(tabs)/user/flight-detail/[id].tsx
@@ -1,10 +1,10 @@
-import Header from "@/conponents/Header";
+import Header from "@/components/Header";
 import { useEffect, useRef, useState } from "react";
 import { Dimensions, Pressable, StyleSheet, Text, View } from "react-native";
 import WebView, { WebViewMessageEvent } from "react-native-webview";
 import Constants from "expo-constants";
-import Close from "@/conponents/icons/Close";
-import MyReportText from "@/conponents/(tabs)/user/my-flight-records/MyReportText";
+import Close from "@/components/icons/Close";
+import MyReportText from "@/components/(tabs)/user/my-flight-records/MyReportText";
 import { useLocalSearchParams } from "expo-router";
 import { getFlightLog } from "@/store/flightLogStore";
 import { ITrackData } from "@/types";
@@ -22,9 +22,7 @@ export default function MyFlightDetails() {
       try {
         const data = await getFlightLog(`${id}`);
         setFlightLog(data);
-        console.log("[id] flightLog", data);
       } catch (error) {
-        console.error("[id] 데이터 로딩 실패: ", error);
       }
     };
 
@@ -51,85 +49,40 @@ export default function MyFlightDetails() {
   `;
 
   useEffect(() => {
-    console.log(
-      "[RN] useEffect triggered - isReady:",
-      isReady,
-      "flightLog:",
-      flightLog?.length
-    );
     if (isReady && flightLog && flightLog.length > 0) {
-      console.log(
-        "[RN] 🚀 Sending flight data with",
-        flightLog.length,
-        "points"
-      );
-      console.log("[RN] First flightLog point:", flightLog[0]);
-
       const message = {
         type: "SET_FLIGHT",
         track: flightLog,
       };
       webviewRef.current?.postMessage(JSON.stringify(message));
-      console.log("[RN] ✅ postMessage called");
-    } else {
-      console.log(
-        "[RN] ⏳ Waiting... isReady:",
-        isReady,
-        "flightLog length:",
-        flightLog?.length || 0
-      );
     }
   }, [isReady, flightLog]);
 
   const onMessage = (e: WebViewMessageEvent) => {
-    console.log("[RN] 📨 Message received from WebView");
-    console.log("[RN] Raw data:", e.nativeEvent.data);
-
     try {
       const raw = e.nativeEvent.data;
       const data = JSON.parse(raw);
-      console.log("[RN] Parsed message type:", data.type);
 
       if (!data?.type) {
-        console.log("[RN] ⚠️ No type in message");
         return;
       }
 
       if (data.type === "READY") {
-        console.log("[RN] ✅ WebView READY received");
         setIsReady(true);
       }
-
-      if (data.type === "PLAY_STARTED") {
-        console.log("[RN] ▶️ play started");
-      }
-      if (data.type === "ERROR") {
-        console.warn("[RN] ❌ webview error:", data.message);
-      }
     } catch (err) {
-      console.error("[RN] 💥 invalid message from webview", err);
     }
   };
 
   const onLoadEnd = () => {
-    console.log("[RN] 🔄 WebView onLoadEnd called");
   };
 
   const onLoadStart = () => {
-    console.log("[RN] 🔄 WebView onLoadStart called");
   };
 
   const onError = (syntheticEvent: any) => {
     const { nativeEvent } = syntheticEvent;
-    console.error("[RN] 💥 WebView error:", nativeEvent);
   };
-
-  console.log(
-    "[RN] 🔄 Component rendering - flightLog:",
-    flightLog?.length,
-    "isReady:",
-    isReady
-  );
 
   return (
     <View style={{ flex: 1 }}>
@@ -169,7 +122,6 @@ export default function MyFlightDetails() {
           <View style={styles.overlay}>
             <Pressable
               onPress={() => {
-                console.log("onClose!");
                 setVisible(false);
               }}
               style={styles.reportCloseBtn}
