+++ libs/(tabs)/air/extractTourList.ts
@@ -2,7 +2,6 @@ import { TourismSchedule } from "@/types";
 import { Alert } from "react-native";
 
 export default function extractTourList(schedules: TourismSchedule[]) {
-  console.log("[extractTourList] ", schedules);
   try {
     const result = [];
     const uniqueSchedules = schedules.filter(
@@ -34,7 +33,6 @@ export default function extractTourList(schedules: TourismSchedule[]) {
 
     return result;
   } catch (error) {
-    console.error("여행 일정 불러오기 오류: ", error);
     Alert.alert("여행 일정을 불러오기 중 오류가 발생했습니다!");
     return [];
   }
