+++ libs/(tabs)/explore/map/fetchSpotMarkers.ts
@@ -17,7 +17,6 @@ export async function fetchSpotMarkers(request: MarkerSpotApiRequest): Promise<M
     });
     return response.data.searchedSpots;
   } catch (error) {
-    console.log("❌ 패러글라이딩 스팟 조회 실패:", error);
     return []; // 실패 시 기본값 반환
   }
 }
